+++ NEWS.md
@@ -1,4 +1,11 @@
-## qtl2 0.7-2 (2017-12-03
+## qtl2 0.7-3 (2017-12-04)
+
+### New features
+
+- The output of `fit1()` now includes fitted values.
+
+- Added function `pull_genoprobpos()` for pulling out a specific
+  position (by name or position) from a set of genotype probabilities.
 
 ### Bug fixes